+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # NEXT RELEASE
-* Add `forEach()` to `CountedList` class
+* Update filter option counts for followed hashtags with substring search even if there's no "#"
 * Turn `TAG_ONLY_STRINGS` into a `config.toots.tagOnlyStrings` parameter; add "is" to list
+* Add `forEach()` to `CountedList` class
 
 ### v0.73.14
 * Keep track of which Toots have had their foreign script hashtags suppressed from the filter options